Add list filters
The drop point list is now a table and has some buttons on top that allow
filtering the rows. The first two buttons do not yet work but later are
intended for filtering drop points based on location or state. The third
button already contains an option to switch display of removed drop points.
The color of the button indicates wether filters are active or not.

The CSS file has been renamed to c3bottles.css to avoid confusion. The CSS
disables transitions for the rows filtered out to make that look less broken.
Some Javascript has been added for the fancy dropdown button state indication.