Vastly improve map logic
Every drop point now has its own layer in the map and can be manipulated
individually. A reference to the layer is saved in the drop point object.
When the state of a drop point changes, this single marker can be easily
redrawn using redraw_marker() instead of redrawing the layer with all
markers.