Add client-side priority calculation
The API now gives the base time and factor for priority calculation and enables
the client to calculate the visit priority locally. The list contains a timer
that redraws the table periodically, refrehsing the priorities displayed.

This fixes #18. However, we now have #20 since server-calculated and
client-calculated priorities are displayed and there might be a difference if
the client's system clock is running at an offset to the server's. In addition,
the priority displayed in the modal on the map is not yet refreshed
automatically (#21).