Pass URL parameters in login redirects
The login form now contains another hidden field for the URL parameters to pass
on for the redirect back to the page from which the user was logging in.
Parsing the passed string needs some hacking since it can be manipulated and
needs to be converted into a dict safely.

Closes: #25