Add OpenStreetMap map source
It is now possible to switch to OSM in the `map.js` file by chaning the
layer. Now, by default no layer is active and the user has to choose one.
Maybe I will make that configurable somehow later.

However, layers and rooms are still hardcoded to the CCH, so this will
break automatic filling of the location as well as drop point numbering
for now.

As that is a different issue (#60), this should close #68 for now.