Make Prometheus metrics configurable
The Prometheus listen address and port are now configurable. The settings
are available in config.default.py. By default, Prometheus is disabled.
If it is enabled, it will listen on 127.0.0.1 and port 9567 by default.

Prometheus will be enabled only if c3bottles is started using WSGI, not
in the development web server. INSTALL.md now has a few lines on Prometheus.

In addition, the module search path modifications in wsgi.py have been
commented out as they are only needed in ancient setups and the structure
for imports from lib has been made more concise with the rest.