Change Dockerfile to better cache build steps
The Dockerfile now has different steps for installing the system dependencies,
installing the Python dependencies and installing the JavaScript dependencies.
Only after that, the actual frontend build takes place.

This allows for quick rebuilds of the image on code changes as we first copy
requirements.txt and package.json step by step, so the Python venv will only
be rebuilt when requirements.txt changes and the Node dependencies are only
installed again if either the Python venv changes (which is the most
time-consuming step of the build process anyway, so the JS stuff doesn't
really matter here) or package.json/yarn.lock change.

entrypoint.sh is not needed if we call Gunicorn directly from within the venv.