Restructure views and templates
All the views are now blueprints which splits up things in a more
natural way. All the view functions and templates have been renamed
or moved to be consistent with the view structure.

Now each template file has the same name as the view function and
is stored in a directory with the same name as the blueprint.

In addition, the label style is now configurable.