Prevent loading the configuration when testing
`c3bottles/__init__.py` loads the configuration on top. Of course, this is
done as well when the tests are run and there might be some stuff happening
during module loading before the test setup changes the configuration.

Therefore, the easiest way to be sure is to not load the configuration at all
when testing. The easiest way to do this is to simply check if the pytest
module is loaded. If not, load the config, if yes, skip loading the config
as pytest will do it later.

Closes #116