Put drop point state into drop point table
This vastly increases the performance of queries for the drop point state
by a factor of over 30. Before, a query for the sate of drop point needed
three database queries (for every single drop point whose state is needed)
whereas now, only a single query can be used for all of them including
their states.