Fix tests and remove disabling of autoflush
Autoflush was once disabled to deal with #65 but actually the issue here
is that a location was added for a drop point before the drop point
cration was finished. If the drop point had errors, the creation of
the location would trigger the flusing of the session.

This caused the problem described in #65 and was dealt with by disabling
autoflusing while failing ealier in drop point creation would have solved
the issue as well.

Now I added explicit DB commits in the location constructor and thereby
got a regression of the issue. This can be easily solved by raising an
exception before constructing a location if drop point creation would
have failed anyway. When deauling with the problem this way, autoflush
can be re-enabled again safely (at least all tests are green now).