Rework API, fix CSRF bug
The API code has been polished and looks a lot better now.

The JS now polls the API regularly to refresh the CSRF token. This ensures
that actions like reporting/visiting will work when the site stays open
for a long time (longer than the CSRF token lifetime).

In addition, the endpoint to poll the latest updates has been moved from
POST to GET which helps with caching and is the right way anyway as it
does not change anything.

Closes #130